+++ Scripts/install.sh
@@ -22,9 +22,16 @@ swift package clean
 
 SWIFT_VERSION=`swift -version | head -n 1 | perl -lpe 's/.*version\ (\d\.\d\.\d).*/$1/'`
 
+UNAME_OUT="$(uname -s)"
+
+case "${UNAME_OUT}" in
+    Darwin*)    IS_MAC=true;;
+    *)          IS_MAC=false;;
+esac
+
 lower $SWIFT_VERSION '5.3.0'
 
-if [[ "$?" -eq 0 ]]
+if [[ "$?" -eq 0 || $IS_MAC ]]
 then
     BUILD_FOLDER=".build/release"
     swift build --disable-sandbox -c release